migrations: fixed data type & inconsistencies seeds: -implemented reservation seeds

+++ server/database/migrations/create_return.sql
@@ -4,7 +4,7 @@ CREATE TABLE Returns (
     time        time,
     odometer    int,
     fulltank    int, 
-    value       int,
+    value       float,
     PRIMARY KEY (rid),
     FOREIGN KEY (rid) REFERENCES Rent
 );
